# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: vercel-postgres-trial
  version: 1.0.0
servers: 
  - url: http://localhost:3000
    description: 開発環境
...
paths:
...
  /api/othello/{gameId}:
    put:
      tags:
        - othello
      parameters:
        - name: gameId
          in: path
          description: game id
          required: true
          schema:
            type: number
      requestBody:
        content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
        message:
          type: string